// src/modules/invoice/usecase/find-invoice/find-invoice.dto.ts
export interface IFindInvoiceUseCaseInputDTO {
    id: string;
}
  
export interface IFindInvoiceUseCaseOutputDTO {
    id: string;
    name: string;
    document: string;
    address: {
        street: string;
        number: number;
        complement: string;
        city: string;
        state: string;
        zipCode: string;
    };
    items: {
        id: string;
        name: string;
        price: number;
    }[];
    total: number;
    createdAt: Date;
}